Wrote new query parser. Bodjed query solver code
I wrote a new query parser, to solve issue #3, and bodged the query
solving code in the process. The query parsing code has been *kinda*
patched. The query parsing code can be found in the `parse_query`
function.

If there is an error with it, please either fix it or inform me.